device-model extracted to own file
