+++ common/ux-window.cpp
@@ -5,10 +5,13 @@
 #define NOMINMAX
 #endif
 #include <glad/glad.h>
-
+#include <rs-config.h>
 #include "ux-window.h"
 
-#include "model-views.h"
+#include <imgui.h>
+#include <imgui_impl_glfw.h>
+
+#include "device-model.h"
 #include "os.h"
 
 // We use STB image to load the splash-screen from memory
